+++ src/Form/types/ObjectInput/Multiple/index.js
@@ -1,0 +1,3 @@
+import Multiple from './Multiple';
+
+export default Multiple;
