+++ src/Form/types/ObjectInput/Single/index.js
@@ -1,0 +1,3 @@
+import TextObject from './Single';
+
+export default TextObject;
