+++ src/Form/types/ObjectInput/index.js
@@ -1,0 +1,2 @@
+export { default as Single } from './Single';
+export { default as Multiple } from './Multiple';
