--- src/Form/types/TextMultipleObject/helpers.js
@@ -1,4 +1,0 @@
-export const clearObjectValues = (obj) => {
-  const objectEmptyValues = Object.keys(obj).map((key) => ({ [key]: '' }));
-  return Object.assign({}, ...objectEmptyValues);
-};
