--- src/Form/types/TextMultipleObject/index.js
@@ -1,3 +1,0 @@
-import TextMultipleObject from './TextMultipleObject';
-
-export default TextMultipleObject;
