--- src/Form/types/TextObject/index.js
@@ -1,3 +1,0 @@
-import TextObject from './TextObject'
-
-export default TextObject;
